+++ CHANGES.md
@@ -12,6 +12,7 @@ Changes in CUPS v2.4.8 (TBA)
 - Updated IPP Everywhere printer creation error reporting (Issue #347)
 - Raised `cups_enum_dests()` timeout for listing available IPP printers
   (Issue #751)
+- Now report an error for temporary printer defaults with lpadmin (Issue #237)
 - Fixed mapping of PPD InputSlot, MediaType, and OutputBin values (Issue #238)
 - Fixed "document-unprintable-error" handling (Issue #391)
 - Fixed the web interface not showing an error for a non-existent printer